requirements.txt
paho-mqtt
requests
pydantic
pydantic-settings
numpy
pandas
pandas-ta
pyyaml
